requirements.txt
# Yêu cầu cơ bản của Python (>= 3.9)

# 1. Bất đồng bộ, Môi trường & Web Server
aiohttp==3.9.3
aiohttp-cors==0.7.0
pydantic==2.6.3
pydantic-settings==2.2.1

# 2. Giao tiếp phần cứng
pyserial==3.5

# 3. Xử lý ảnh & AI
numpy>=1.24.0
# Dùng opencv-python-headless trên Pi để tránh lỗi thiếu thư viện giao diện (GUI X11)
opencv-python-headless==4.9.0.80
face_recognition==1.3.0
# Ghi chú: picamera2 thường được cài sẵn trên Raspberry Pi OS bản mới.
# Nếu thiếu, cài qua lệnh hệ thống: sudo apt install python3-picamera2

# 4. Cloud Storage (Tùy chọn theo backend bạn dùng)
boto3==1.34.51        # Nếu dùng AWS S3
cloudinary==1.39.0    # Nếu dùng Cloudinary
